Finalising the main function

+++ db.py
@@ -1,11 +1,12 @@
 # This model for writing and reading money
 import csv
+from decimal import Decimal
 def read_money():
     try:
         with open('money.txt','r') as file:
             reader = csv.reader(file)
             money = next(reader)
-            return float(money[0])
+            return Decimal(money[0])
     except FileNotFoundError:
         print("Money file not found.Creating new one")
         write_money(100)
